python_ocr_server/requirements.txt
# FastAPI and web server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PaddleOCR and dependencies
paddlepaddle==3.0.0b2
paddleocr==2.9.1

# Image processing
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.3.5

# Utilities
pydantic==2.12.4
